+++ src/config.rs
@@ -30,15 +30,8 @@ pub struct NodeConfig {
     /// note: this number can very briefly be breached by 1 in case of inbound connection attempts. It can never be
     /// breached by outbound connection attempts, though.
     pub max_connections: u16,
-    /// The maximum time allowed for connections to enable all protocols.
-    ///
-    /// note: it should not be treated as a handshake timeout; the handshake is primarily intended to be an operation
-    /// synchronous towards the `Connection` object: if a `Connection` is sent to the `Handshaking` protocol handler
-    /// and the associated task hangs, it will not automatically get dropped and the task will not resume - it is the
-    /// user's responsibility to handle it. The same warning applies to `Reading` and `Writing`, but they are intended
-    /// to immediately spawn their own tasks (as in the default implementations), so they are less susceptible to this
-    /// issue.
-    pub max_protocol_setup_time_ms: u64,
+    /// The maximum time allowed for a connection to perform a handshake before it is rejected.
+    pub max_handshake_time_ms: u64,
 }
 
 impl Default for NodeConfig {
@@ -55,7 +48,7 @@ impl Default for NodeConfig {
             conn_outbound_queue_depth: 16,
             invalid_read_delay_secs: 10,
             max_connections: 100,
-            max_protocol_setup_time_ms: 3000,
+            max_handshake_time_ms: 3000,
         }
     }
 }